GCW-3675 Encapsulate typing activity in hook
